requirements.txt
# Discord-bot core
discord.py==2.6.4
python-dotenv==1.2.1

# Logging & färgade terminalutskrifter
colorama==0.4.6
loguru==0.7.2

# HTTP & API-anrop (för AI eller externa tjänster)
requests==2.31.0
aiohttp==3.9.7
httpx==0.25.0

# JSON & datahantering
ujson==5.8.0

# Schemaläggning & bakgrundsjobb
apscheduler==3.10.4

# Extra verktyg för stabil drift
psutil==5.9.8
rich==13.7.0

# Voice support
PyNaCl==1.5.0